+++ update.ubuntu.sh
@@ -18,7 +18,7 @@
 # DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-version="1.6.5";
+version="2.0.0";
 
 tput setaf 2;
 echo "Check: root user;";
@@ -87,24 +87,24 @@ set -eu;
 cd ~/$misskey_directory;
 
 tput setaf 3;
-echo "Process: yarn install and clean;";
-tput setaf 7;
-npx yarn install;
-
-tput setaf 3;
 echo "Process: clean;";
 tput setaf 7;
-npm run clean;
+yarn run clean;
+
+tput setaf 3;
+echo "Process: yarn install;";
+tput setaf 7;
+NODE_ENV=production yarn install --immutable;
 
 tput setaf 3;
 echo "Process: build misskey;";
 tput setaf 7;
-NODE_OPTIONS=--max_old_space_size=3072 NODE_ENV=production npm run build;
+NODE_OPTIONS=--max_old_space_size=3072 NODE_ENV=production yarn run build;
 
 tput setaf 3;
 echo "Process: migrate db;";
 tput setaf 7;
-NODE_OPTIONS=--max_old_space_size=3072 npm run migrate;
+NODE_OPTIONS=--max_old_space_size=3072 yarn run migrate;
 MKEOF
 #endregion